webservice/requirements.txt
blinker==1.6.3
click==8.1.7
Flask==3.0.0
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
MarkupSafe==2.1.3
packaging==23.2
Werkzeug==3.0.0
